ResponseEntity Implemented on CURD

+++ src/main/java/org/rajbhar/journalapp/controller/JournalEntryControllerV2.java
@@ -5,10 +5,13 @@ import org.bson.types.ObjectId;
 import org.rajbhar.journalapp.entity.JournalEntry;
 import org.rajbhar.journalapp.service.JournalEntryService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/journal")                 //entry point url -> /journal/xxxx
@@ -19,39 +22,62 @@ public class JournalEntryControllerV2 {
 
     //localhost:8080/journal GET
     @GetMapping                               //(<empty>) -> '/'
-    public List<JournalEntry>getAll(){          //method  //it should always be public
-        return journalEntryService.getAll();
+    public ResponseEntity<?>getAll(){          //method  //it should always be public
+        //return journalEntryService.getAll();
+        List<JournalEntry> all =  journalEntryService.getAll();
+        if(all != null && !all.isEmpty()){
+            return new ResponseEntity<>(all, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     //localhost:8080/journal POST
     @PostMapping
-    public JournalEntry createEntry(@RequestBody JournalEntry myEntry){    //body,type, any name
-        myEntry.setDate(LocalDateTime.now());
-        journalEntryService.saveEntry( myEntry);
-        return myEntry;
+    public ResponseEntity<JournalEntry> createEntry(@RequestBody JournalEntry myEntry){    //body,type, any name
+        try{
+            myEntry.setDate(LocalDateTime.now());
+            journalEntryService.saveEntry( myEntry);
+            //return myEntry;
+            return new ResponseEntity<>(myEntry, HttpStatus.CREATED);
+
+        }catch(Exception e){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
 
     @GetMapping("/id/{myId}")
-    public JournalEntry getJournalEntryById(@PathVariable ObjectId myId){
-        return journalEntryService.findById(myId).orElse(null);
+    public ResponseEntity<JournalEntry> getJournalEntryById(@PathVariable ObjectId myId){
+        //return journalEntryService.findById(myId).orElse(null);
+        Optional<JournalEntry> journalEntry = journalEntryService.findById(myId);
+        if(journalEntry.isPresent()){
+            return new ResponseEntity<>(journalEntry.get(), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
+
     }
 
     @DeleteMapping("/id/{myId}")
-    public boolean deleteJournalEntryById(@PathVariable ObjectId myId){
-        journalEntryService.deleteById(myId);
-        return true;
+    public ResponseEntity<?> deleteJournalEntryById(@PathVariable ObjectId myId){
+        JournalEntry journalEntry = journalEntryService.findById(myId).orElse(null);
+        if(journalEntry != null){
+            journalEntryService.deleteById(myId);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @PutMapping("/id/{myId}")
-    public JournalEntry updateJournalEntryById(@PathVariable ObjectId myId, @RequestBody JournalEntry newEntry){   //myId = old, myEntry = new
+    public ResponseEntity<?> updateJournalEntryById(@PathVariable ObjectId myId, @RequestBody JournalEntry newEntry){   //myId = old, myEntry = new
         JournalEntry oldEntry = journalEntryService.findById(myId).orElse(null);
         if(oldEntry != null){
-            oldEntry.setTitle(newEntry.getTitle() != null && !newEntry.getTitle().equals("") ? newEntry.getTitle() :  oldEntry.getTitle());
-            oldEntry.setContent(newEntry.getContent() != null && !newEntry.getContent().equals("") ? newEntry.getContent() : oldEntry.getContent());
+            oldEntry.setTitle(newEntry.getTitle() != null && !newEntry.getTitle().isEmpty() ? newEntry.getTitle() :  oldEntry.getTitle());
+            oldEntry.setContent(newEntry.getContent() != null && !newEntry.getContent().isEmpty() ? newEntry.getContent() : oldEntry.getContent());
+            journalEntryService.saveEntry(oldEntry);
+            //return oldEntry;
+            return new ResponseEntity<>(oldEntry, HttpStatus.OK);
         }
-
-        journalEntryService.saveEntry(oldEntry);
-        return oldEntry;
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
 }
